apps/api/requirements.txt
# apps/api/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-filter==23.5
python-dotenv==1.0.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.27